QString replacement builds, runs and pops up and error (this is likely to have pushed many regressions)

+++ src/ta3d/src/ingame/menus/statistics.h
@@ -64,7 +64,7 @@ namespace Menus
 		** \param color The color of the player
 		** \param value The text
 		*/
-		void doUpdateObject(const QString& id, const short indx, const uint32 color, const QString& value);
+        void doUpdateObject(const QString& id, const int indx, const uint32 color, const QString& value);
 
 	}; // class Statistics
